fixed f32 error

+++ core/src/utils/services/llm.rs
@@ -92,8 +92,10 @@ async fn open_ai_call_handler(
             ctx.span()
                 .set_attribute(opentelemetry::KeyValue::new("rate_limited", true));
             if let Some(rate_limit) = model.rate_limit {
-                ctx.span()
-                    .set_attribute(opentelemetry::KeyValue::new("rate_limit", rate_limit));
+                ctx.span().set_attribute(opentelemetry::KeyValue::new(
+                    "rate_limit",
+                    rate_limit as f64,
+                ));
             }
             rate_limiter
                 .acquire_token_with_timeout(std::time::Duration::from_secs(
